Right-size types based on analysis of all games.

+++ src/bgg/thing/thing1.rs
@@ -53,7 +53,7 @@ pub(super) struct Name {
 // example: <minplayers value="2"/>
 #[derive(Deserialize)]
 pub(super) struct ElementWithIntValueAttribute {
-    pub(super) value: u8,
+    pub(super) value: u16,
 }
 
 #[derive(Deserialize)]
